pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "airsoft-suitcase"
version = "0.1.0"
description = "Interactive airsoft game system for Raspberry Pi"
requires-python = ">=3.9"
dependencies = [
    "pygame>=2.6",
    "pynput>=1.8",
]

[project.optional-dependencies]
pi = [
    "RPi.GPIO>=0.7",
    "adafruit-blinka>=8.0",
    "adafruit-circuitpython-neopixel>=6.3",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.9",
]

[tool.ruff]
target-version = "py39"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]
ignore = ["UP006", "UP035"]

[tool.ruff.lint.per-file-ignores]
"scripts/*" = ["E402", "I001"]

[tool.pytest.ini_options]
testpaths = ["tests"]
